+++ netbox_contract/__init__.py
@@ -18,6 +18,15 @@ class ContractsConfig(PluginConfig):
         'mandatory_invoice_fields': [],
         'hidden_invoice_fields': [],
         'mandatory_dimensions': [],
+        'supported_models': [
+            'circuits.circuit',
+            'circuits.virtualcircuit',
+            'dcim.site',
+            'dcim.device',
+            'dcim.rack',
+            'virtualization.virtualmachine',
+            'virtualization.cluster',
+        ],
     }