+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Version 2.4.3
 
 * [282](https://github.com/mlebreuil/netbox-contract/issues/282) Add the possibility to assign contract to clusters and racks.
+* [276](https://github.com/mlebreuil/netbox-contract/issues/276) Add the possibility to update the provider in bulk for contracts.
 * [275](https://github.com/mlebreuil/netbox-contract/issues/275) Add the possibility to filter contract by provider an service provider. Remove the default value for the currency and internal party fields in the contract search form.
 * [274](https://github.com/mlebreuil/netbox-contract/issues/274) List contracts on providers and service providers pages